Subida automática

+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: bazuara <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 12:49:44 by bazuara           #+#    #+#             */
-/*   Updated: 2019/11/06 13:10:01 by bazuara          ###   ########.fr       */
+/*   Updated: 2019/11/12 10:31:36 by bazuara          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s1[i] != '\0' && s2[i] != '\0' && s1[i] == s2[i] && i < n)
